+++ bricks/debug/mpconfigport.h
@@ -36,6 +36,7 @@
 #define PYBRICKS_PY_PUPDEVICES                  (0)
 #define PYBRICKS_PY_ROBOTICS                    (0)
 #define PYBRICKS_PY_TOOLS                       (1)
+#define PYBRICKS_PY_TOOLS_HUB_MENU              (0)
 
 // Pybricks options
 #define PYBRICKS_OPT_COMPILER                   (0)