+++ bricks/essentialhub/mpconfigport.h
@@ -42,6 +42,7 @@
 #define PYBRICKS_PY_ROBOTICS_DRIVEBASE_GYRO     (1)
 #define PYBRICKS_PY_ROBOTICS_DRIVEBASE_SPIKE    (1)
 #define PYBRICKS_PY_TOOLS                       (1)
+#define PYBRICKS_PY_TOOLS_HUB_MENU              (0)
 
 // Pybricks options
 #define PYBRICKS_OPT_COMPILER                   (1)