+++ bricks/ev3dev/brickconfig.h
@@ -44,4 +44,5 @@
 #define PYBRICKS_PY_ROBOTICS            (1)
 #define PYBRICKS_PY_ROBOTICS_DRIVEBASE_SPIKE (0)
 #define PYBRICKS_PY_TOOLS               (1)
+#define PYBRICKS_PY_TOOLS_HUB_MENU      (0)
 #define PYBRICKS_PY_USIGNAL             (1)