+++ CHANGELOG.md
@@ -8,6 +8,7 @@
 - Added `'modes'` entry to the dictionary returned by `PUPDevice.info()`. It
   is a tuple of `(name, num_values, data_type)` tuples for each available mode.
 - Added `pybricks.tools.read_input_byte()` function ([support#1102]).
+- Added `pybricks.tools.hub_menu()` function ([support#1064]).
 
 ### Changed
 - Changed internal drivers for LEGO devices (motors and sensors) on all platforms.
@@ -16,6 +17,7 @@
 - Fixed hub will not power off when Bluetooth chip crashes on City and Technic hubs ([support#1095]).
 - Fixed `off()` method in `ColorLightMatrix`, `UltrasonicSensor`, `ColorSensor` ([support#1098]).
 
+[support#1064]: https://github.com/pybricks/support/issues/1064
 [support#1095]: https://github.com/pybricks/support/issues/1095
 [support#1098]: https://github.com/pybricks/support/issues/1098
 [support#1102]: https://github.com/pybricks/support/issues/1102